requirements.txt
Flask
fastmcp
pytesseract
Pillow
python-dotenv
google-generativeai
langchain-google-genai
langgraph
langchain-mcp-adapters
